makes type overrideable and hint format from type.
Introduces a new field in the JsonFormat annotation to explicitly set the type
(e.g. if a serializer writes a string for a numeric field).

Introduces a set of format hints for a number of data types.

Note: uses non-standard "uuid" for UUID fields, assuming that the
standards committee gets their act together and adds that post-draft 8
to the spec.